(svn r1496) -Fix: highscore no longer crashes in network games with a dedicated server. At the end of the game (can only be set by the server) the highscore is shown for the top5 companies of that game -Fix: fixed some compiler warnings -Added PF_NETWORK_ONLY flag to settings. Such a setting can only be modified in a network game.

+++ functions.h
@@ -234,7 +234,7 @@ void ShowNetworkNeedGamePassword();
 void ShowNetworkNeedCompanyPassword();
 void ShowRenameWaypointWindow(Waypoint *cp);
 int FindFirstBit(uint32 x);
-void ShowHighscoreTable(int difficulty, int rank);
+void ShowHighscoreTable(int difficulty, int8 rank);
 void ShowEndGameChart(void);
 TileIndex AdjustTileCoordRandomly(TileIndex a, byte rng);
